requirements.txt
customtkinter==5.2.2
PyPDF2==3.0.1
pillow==10.2.0 